lab/app-for-protocol-tests/pyproject.toml
[tool.poetry]
name = "app-for-protocol-tests"
version = "0.1.0"
description = ""
authors = ["Johnny Tardin <johnnytardin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
postgres = "^4.0"
Flask = "^2.0.2"
gevent = "^21.8.0"
requests = "^2.26.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
